+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added integration testing for the lib
+- Browser package has returned using HypheLib()
 
 ### Fixed